ide-module/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
jinja2>=3.0.1
pydantic>=1.8.2
docker>=5.0.3
python-dotenv>=0.19.0